fixed warnings; added libgcc to the cross compiler; removed float functions from kernel space

+++ kernel/arch/x86/context.c
@@ -50,7 +50,7 @@ struct cpu_state *arch_create_thread_context(struct arch_thread_context *context
     void *kernel_stack = malloc(0x1000);
     struct cpu_state *new_state = kernel_stack + 0x1000 - sizeof(struct cpu_state) - 3*sizeof(uint32_t);
 
-	context->kernel_mode_stack = kernel_stack;
+	context->kernel_mode_stack = (uint32_t)kernel_stack;
     context->state = new_state;
 	
     memset(new_state, 0, sizeof(struct cpu_state));
